# Conflicts:
#	gui/admin-gui/pom.xml
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/api/page/PageBase.html
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/api/page/PageBase.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/api/util/WebModelServiceUtils.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/component/ContainerableListPanel.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/component/menu/LeftMenuPanel.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/page/admin/AbstractPageObjectDetails.html
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/page/login/PageLoginOld.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/boot/AbstractSpringBootApplication.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/admin/configuration/PageAbout.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/admin/home/PageDashboard.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/forgetpassword/PageResetPassword.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/forgetpassword/PageSecurityQuestions.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/login/AbstractPageLogin.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/login/PageAuthenticationBase.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/login/PageSamlSelect.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/page/self/PagePostAuthentication.java
#	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/security/util/SecurityUtils.java
#	gui/admin-gui/src/main/resources/static/js/midpoint-theme.js
#	gui/admin-gui/src/main/resources/static/js/midpoint/ace-editor.js

+++ repo/repo-test-util/src/main/java/com/evolveum/midpoint/test/TaskFinishChecker.java
@@ -34,7 +34,7 @@ public class TaskFinishChecker implements Checker {
     private final OperationResult waitResult;
     private final boolean checkSubresult;
     private final boolean errorOk;
-    private final int timeout;
+    private final long timeout;
     private final int showProgressEach;
     private final boolean verbose;
     private final Consumer<Task> taskConsumer;
@@ -123,7 +123,7 @@ public class TaskFinishChecker implements Checker {
         private OperationResult waitResult;
         private boolean checkSubresult;
         private boolean errorOk;
-        private int timeout;
+        private long timeout;
         private int showProgressEach;
         private boolean verbose;
         private Consumer<Task> taskConsumer;
@@ -159,7 +159,7 @@ public class TaskFinishChecker implements Checker {
             return this;
         }
 
-        public Builder timeout(int val) {
+        public Builder timeout(long val) {
             timeout = val;
             return this;
         }
